Refactored typeDefs and resolvers to their own directories for a more clean implementation, added nodemon dev dependency
