+++ graphql/resolvers/index.js
@@ -1,0 +1,8 @@
+const postResolvers = require('./posts');
+const userResolvers = require('./users');
+
+module.exports = {
+  Query: {
+    ...postResolvers.Query,
+  },
+};
