+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "nodemon index"
   },
   "keywords": [],
   "author": "",
@@ -13,5 +13,8 @@
     "apollo-server": "^2.19.0",
     "graphql": "^15.4.0",
     "mongoose": "^5.11.4"
+  },
+  "devDependencies": {
+    "nodemon": "^2.0.6"
   }
 }
